refactor(coding-agent): extract commands and tools from AcpSession
